added loading skeleton to avatar, fixed redirects for auth and within /app pages, swapped hard-coded profileId to supabase's profile_id
