+++ apps/web/src/lib/components/ui/coming-soon/coming-soon-page.svelte
@@ -17,14 +17,3 @@
 		</p>
 	</div>
 </main>
-
-<style>
-	h1 {
-		font-size: 2rem;
-		margin-bottom: 1rem;
-	}
-
-	p {
-		font-size: 1.2rem;
-	}
-</style>
