+++ apps/web/src/lib/schema.ts
@@ -15,4 +15,19 @@ export const formSchema = z.object({
 		.regex(/[0-9]/, { message: "Password must contain at least one number." })
 });
 
+export const loginUserSchema = z.object({
+	email: z.string().email({ message: "Invalid email address" }),
+	password: z
+		.string()
+		.min(8, { message: "Password must be at least 8 characters long." })
+		.max(100, { message: "Password must be 100 characters or less." })
+		.regex(/[a-z]/, { message: "Password must contain at least one lowercase letter." })
+		.regex(/[A-Z]/, { message: "Password must contain at least one uppercase letter." })
+		.regex(/[0-9]/, { message: "Password must contain at least one number." })
+});
+
+export const waitlistSchema = z.object({
+	email: z.string().email({ message: "Invalid email address" })
+});
+
 export type FormSchema = typeof formSchema;
