+++ apps/web/src/routes/(marketing)/(auth)/+layout.svelte
@@ -1,0 +1,5 @@
+<div class="max-w-screen mx-auto flex h-screen items-center justify-center">
+	<div class="w-full max-w-lg rounded-lg">
+		<slot />
+	</div>
+</div>
