+++ apps/web/src/routes/(marketing)/(auth)/login/+page.server.ts
@@ -3,17 +3,6 @@ import type { Actions, PageServerLoad } from "./$types";
 import type { Provider } from "@supabase/supabase-js";
 import { z } from "zod";
 
-const loginUserSchema = z.object({
-	email: z.string().email({ message: "Invalid email address" }),
-	password: z
-		.string()
-		.min(8, { message: "Password must be at least 8 characters long." })
-		.max(100, { message: "Password must be 100 characters or less." })
-		.regex(/[a-z]/, { message: "Password must contain at least one lowercase letter." })
-		.regex(/[A-Z]/, { message: "Password must contain at least one uppercase letter." })
-		.regex(/[0-9]/, { message: "Password must contain at least one number." })
-});
-
 export const load = (async () => {
 	return {};
 }) satisfies PageServerLoad;
@@ -54,6 +43,6 @@ export const actions: Actions = {
 			});
 		}
 
-		throw redirect(307, "/");
+		throw redirect(307, "/app");
 	}
 };
