+++ apps/web/src/routes/app/+page.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+	import { goto } from '$app/navigation';
+	import { onMount } from 'svelte';
+
+	onMount(() => {
+		goto('/app/editor/crew');
+	});
+</script>
