+++ apps/web/src/routes/app/auto-build/+page.svelte
@@ -2,4 +2,4 @@
 	import ComingSoonPage from '$lib/components/ui/coming-soon/coming-soon-page.svelte';
 </script>
 
-<ComingSoonPage releaseVersion="v0.2.2" />
+<ComingSoonPage releaseVersion="v0.2.0" />
