+++ apps/web/src/routes/app/library/agent/+page.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+	import { ComingSoonPage } from '$lib/components/ui/coming-soon';
+</script>
+
+<ComingSoonPage releaseVersion="v0.2.0" />
