+++ apps/web/src/routes/app/library/crew/+page.svelte
@@ -1,0 +1,5 @@
+<script>
+	import { ComingSoonPage } from '$lib/components/ui/coming-soon';
+</script>
+
+<ComingSoonPage releaseVersion="v0.2.0" />
